requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
email-validator==2.1.0
apscheduler==3.10.4
jinja2==3.1.2
resend==0.3.0
yfinance==0.2.32
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
lxml==4.9.3
pandas==2.1.3
numpy==1.26.2
gunicorn==21.2.0
